+++ crates/chainio/clients/avsregistry/src/writer.rs
@@ -459,6 +459,30 @@ impl AvsRegistryChainWriter {
         Ok(*tx.tx_hash())
     }
 
+    /// Set the ejection cooldow that an operator must wait in seconds after ejection before registering for any quorum
+    ///
+    /// # Arguments
+    ///
+    /// * `cooldown` - the new ejection cooldown in seconds
+    ///
+    /// # Returns
+    ///
+    /// * `TxHash` - The transaction hash of the set ejection cooldown transaction
+    pub async fn set_ejection_cooldown(&self, cooldown: U256) -> Result<TxHash, AvsRegistryError> {
+        info!("setting ejecting cooldown with the AVS's registry coordinator");
+        let provider = get_signer(&self.signer.clone(), &self.provider);
+
+        let contract_registry_coordinator =
+            RegistryCoordinator::new(self.registry_coordinator_addr, provider);
+
+        contract_registry_coordinator.setEjectionCooldown(cooldown)
+            .send()
+            .await
+            .map_err(AvsRegistryError::AlloyContractError)
+            .inspect(|tx| info!(tx_hash = ?tx, "successfully set ejection cooldown with the AVS's registry coordinator"))
+            .map(|tx| *tx.tx_hash())
+    }
+
     /// Sets an address as the ejector, which can force-deregister operators from quorums.
     ///
     /// # Arguments
@@ -494,6 +518,7 @@ mod tests {
     use crate::test_utils::{
         build_avs_registry_chain_writer, test_deregister_operator, test_register_operator,
     };
+    use alloy::primitives::U256;
     use alloy::primitives::{Address, Bytes};
     use eigen_common::{get_provider, get_signer};
     use eigen_testing_utils::anvil::{start_anvil_container, start_m2_anvil_container};
@@ -761,6 +786,35 @@ mod tests {
             op_params._0.kickBIPsOfTotalStake,
             operator_set_params.kickBIPsOfTotalStake
         );
+    }
+
+    #[tokio::test]
+    async fn test_set_ejector_cooldown() {
+        let (_container, http_endpoint, _ws_endpoint) = start_m2_anvil_container().await;
+        let private_key = FIRST_PRIVATE_KEY.to_string();
+
+        let avs_writer =
+            build_avs_registry_chain_writer(http_endpoint.clone(), private_key.clone()).await;
+
+        let registry_contract = RegistryCoordinator::new(
+            avs_writer.registry_coordinator_addr,
+            get_signer(&avs_writer.signer.clone(), &avs_writer.provider),
+        );
+
+        let new_cooldown = U256::from(100);
+        let tx_hash = avs_writer
+            .set_ejection_cooldown(new_cooldown)
+            .await
+            .unwrap();
+
+        let tx_status = wait_transaction(&http_endpoint, tx_hash)
+            .await
+            .unwrap()
+            .status();
+
+        assert!(tx_status);
+        let cooldown = registry_contract.ejectionCooldown().call().await.unwrap();
+        assert_eq!(cooldown._0, new_cooldown);
     }
 
     #[tokio::test]